Created new Class SaveFile.java
